+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Inline derive macros' docs ([#3])
+
+[#3]: https://github.com/alloy-rs/rlp/pull/3
+
 ## [0.3.1] - 2023-07-13
 
 ### Fixed