+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.3] - 2023-09-23
+
 ### Fixed
 
 - `Ip{,v4,v6}Addr::decode` ([#6])